+++ R/vec-new.R
@@ -147,6 +147,7 @@ ll <- function(...) {
 #' zero-filled vectors.
 #'
 #' @param n The vector length.
+#' @param names Names for the new vector.
 #' @keywords internal
 #' @examples
 #' new_list(10)
@@ -157,38 +158,38 @@ NULL
 
 #' @rdname new-vector
 #' @export
-new_logical <- function(n) {
-  rep_len(na_lgl, n)
+new_logical <- function(n, names = NULL) {
+  set_names(rep_len(na_lgl, n), names)
 }
 #' @rdname new-vector
 #' @export
-new_integer <- function(n) {
-  rep_len(na_int, n)
+new_integer <- function(n, names = NULL) {
+  set_names(rep_len(na_int, n), names)
 }
 #' @rdname new-vector
 #' @export
-new_double <- function(n) {
-  rep_len(na_dbl, n)
+new_double <- function(n, names = NULL) {
+  set_names(rep_len(na_dbl, n), names)
 }
 #' @rdname new-vector
 #' @export
-new_character <- function(n) {
-  rep_len(na_chr, n)
+new_character <- function(n, names = NULL) {
+  set_names(rep_len(na_chr, n), names)
 }
 #' @rdname new-vector
 #' @export
-new_complex <- function(n) {
-  rep_len(na_cpl, n)
+new_complex <- function(n, names = NULL) {
+  set_names(rep_len(na_cpl, n), names)
 }
 #' @rdname new-vector
 #' @export
-new_raw <- function(n) {
-  vector("raw", n)
+new_raw <- function(n, names = NULL) {
+  set_names(vector("raw", n), names)
 }
 #' @rdname new-vector
 #' @export
-new_list <- function(n) {
-  vector("list", n)
+new_list <- function(n, names = NULL) {
+  set_names(vector("list", n), names)
 }
 
 #' Create vectors matching the length of a given vector
@@ -198,53 +199,59 @@ new_list <- function(n) {
 #' Except for `new_list_along()` and `new_raw_along()`, the empty
 #' vectors are filled with typed `missing` values.
 #'
-#' @inheritParams set_attrs
 #' @param x,.x A vector.
 #' @param .y Values to repeat.
+#' @param names Names for the new vector. Defaults to the names of
+#'   `x`. This can be a function to apply to the names of `x` as in
+#'   [set_names()].
 #' @keywords internal
 #' @examples
 #' x <- 0:5
 #' rep_along(x, 1:2)
 #' rep_along(x, 1)
 #' new_list_along(x)
+#'
+#' # The default names are picked up from the input vector
+#' x <- c(a = "foo", b = "bar")
+#' new_character_along(x)
 #' @name new-vector-along
 #' @seealso new-vector
 NULL
 
 #' @export
 #' @rdname new-vector-along
-new_logical_along <- function(x) {
-  rep_len(na_lgl, length(x))
+new_logical_along <- function(x, names = base::names(x)) {
+  set_names_impl(rep_len(na_lgl, length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_integer_along <- function(x) {
-  rep_len(na_int, length(x))
+new_integer_along <- function(x, names = base::names(x)) {
+  set_names_impl(rep_len(na_int, length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_double_along <- function(x) {
-  rep_len(na_dbl, length(x))
+new_double_along <- function(x, names = base::names(x)) {
+  set_names_impl(rep_len(na_dbl, length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_character_along <- function(x) {
-  rep_len(na_chr, length(x))
+new_character_along <- function(x, names = base::names(x)) {
+  set_names_impl(rep_len(na_chr, length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_complex_along <- function(x) {
-  rep_len(na_cpl, length(x))
+new_complex_along <- function(x, names = base::names(x)) {
+  set_names_impl(rep_len(na_cpl, length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_raw_along <- function(x) {
-  vector("raw", length(x))
+new_raw_along <- function(x, names = base::names(x)) {
+  set_names_impl(vector("raw", length(x)), x, names)
 }
 #' @export
 #' @rdname new-vector-along
-new_list_along <- function(x) {
-  vector("list", length(x))
+new_list_along <- function(x, names = base::names(x)) {
+  set_names_impl(vector("list", length(x)), x, names)
 }
 
 #' @export